refactor: сделал небольшие изменения во вспомогательных классах и сервисе.
