+++ src/main/java/team/mediasoft/wareshop/exchanger/webconfig/WebClientConfig.java
@@ -9,7 +9,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.codec.json.Jackson2JsonDecoder;
 import org.springframework.http.codec.json.Jackson2JsonEncoder;
 import org.springframework.web.reactive.function.client.WebClient;
-import team.mediasoft.wareshop.exchanger.util.RestProperties;
+import team.mediasoft.wareshop.util.RestProperties;
 
 @Configuration
 @RequiredArgsConstructor
