+++ src/main/java/team/mediasoft/wareshop/util/RestProperties.java
@@ -1,7 +1,9 @@
-package team.mediasoft.wareshop.exchanger.util;
+package team.mediasoft.wareshop.util;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import java.util.List;
+
 @ConfigurationProperties(prefix = "app.rest.currency-service")
-public record RestProperties(String host, String method, String file) {
+public record RestProperties(String host, List<String> methods, String file) {
 }
